@@ -127,7 +127,7 @@ const Input = {
   focus: DEFAULT_FALSE,
   'confirm-type': singleQuote('done'),
   'confirm-hold': DEFAULT_FALSE,
-  cursor: 'i.value.length',
+  cursor: '-1',
   'selection-start': '-1',
   'selection-end': '-1',
   bindInput: NO_DEFAULT_VALUE,
