@@ -231,13 +231,16 @@ type ITaroHooks = {
   /** 用于修改 Taro DOM 事件对象 */
   modifyTaroEvent: (event, element) => void
 
-  dispatchTaroEvent: (event, element) => void 
+  dispatchTaroEvent: (event, element) => void
   dispatchTaroEventFinish: (event, element) => void
 
   modifyDispatchEvent: (event, element) => void
   injectNewStyleProperties: (styleProperties: string[]) => void
   initNativeApi: (taro: Record<string, any>) => void
   patchElement: (node) => void
+
+  /** 解 Proxy */
+  proxyToRaw: (proxyObj) => Record<any, any>
 }
 
 export const hooks = new TaroHooks<ITaroHooks>({
@@ -298,6 +301,10 @@ export const hooks = new TaroHooks<ITaroHooks>({
 
   onAddEvent: TaroHook(HOOK_TYPE.SINGLE),
 
+  proxyToRaw: TaroHook(HOOK_TYPE.SINGLE, function (proxyObj) {
+    return proxyObj
+  }),
+
   modifyMpEvent: TaroHook(HOOK_TYPE.MULTI),
 
   modifyMpEventImpl: TaroHook(HOOK_TYPE.SINGLE, function (this: TaroHooks<ITaroHooks>, e: MpEvent) {
@@ -312,7 +319,7 @@ export const hooks = new TaroHooks<ITaroHooks>({
   injectNewStyleProperties: TaroHook(HOOK_TYPE.SINGLE),
 
   modifyTaroEvent: TaroHook(HOOK_TYPE.MULTI),
-  
+
   dispatchTaroEvent: TaroHook(HOOK_TYPE.SINGLE, (e, node) => {
     node.dispatchEvent(e)
   }),
