@@ -1,6 +1,9 @@
+import { chalk } from '@tarojs/helper'
 import * as path from 'path'
 
-import validator from '../doctor/eslintValidator'
+import doctor from '../doctor'
+
+const validator = doctor.validators[4]
 
 describe('eslint validator of doctor', () => {
   let cwd = ''
@@ -14,37 +17,41 @@ describe('eslint validator of doctor', () => {
 
   it('should lint for react', async () => {
     process.chdir(path.join(__dirname, 'fixtures/default'))
-    const raw = await validator({
+    const { isValid, messages } = await validator({
       projectConfig: {
         framework: 'react',
         sourceRoot: 'src'
-      }
-    }).then(e => e.raw)
-
-    expect(raw).toBe('')
+      },
+      chalk
+    })
+    expect(isValid).toBe(true)
+    expect(messages.length).toBe(2)
   })
 
   it('should lint for nerv', async () => {
     process.chdir(path.join(__dirname, 'fixtures/nerv'))
-    const raw = await validator({
+    const { isValid, messages } = await validator({
       projectConfig: {
         framework: 'nerv',
         sourceRoot: 'src'
-      }
-    }).then(e => e.raw)
-
-    expect(raw.includes('\'a\' is assigned a value but never used'))
+      },
+      chalk
+    })
+    expect(isValid).toBe(false)
+    expect(messages[1].content.includes('\'a\' is assigned a value but never used'))
   })
 
   it('should lint for vue', async () => {
     process.chdir(path.join(__dirname, 'fixtures/vue'))
-    const raw = await validator({
+    const { isValid, messages } = await validator({
       projectConfig: {
         framework: 'vue',
         sourceRoot: 'src'
-      }
-    }).then(e => e.raw)
+      },
+      chalk
+    })
 
-    expect(raw).toBe('')
+    expect(isValid).toBe(true)
+    expect(messages.length).toBe(2)
   })
 })
