@@ -1,5 +1,4 @@
 import { chalk } from '@tarojs/helper'
-import * as path from 'path'
 
 import { run } from './utils'
 
@@ -8,27 +7,35 @@ jest.mock('../doctor', () => {
     __esModule: true,
     default: {
       validators: [() => ({
-        desc: 'configValidator',
-        lines: [{
-          desc: 'A',
-          valid: false
+        isValid: true,
+        messgaes: [{
+          kind: 2,
+          content: 'Env Success'
         }]
       }), () => ({
-        desc: 'packageValidator',
-        lines: [{
-          desc: 'B',
-          valid: true
-        }, {
-          desc: 'C',
-          valid: true,
-          solution: 'c'
+        isValid: false,
+        messgaes: [{
+          kind: 1,
+          content: 'Config Error'
         }]
       }), () => ({
-        desc: 'recommandValidator',
-        lines: []
+        isValid: true,
+        messgaes: [{
+          kind: 2,
+          content: 'Package Success'
+        }]
       }), () => ({
-        desc: 'eslintValidator',
-        raw: 'eslint msg'
+        isValid: true,
+        messgaes: [{
+          kind: 2,
+          content: 'recommend Success'
+        }]
+      }), async () => ({
+        isValid: true,
+        messgaes: [{
+          kind: 2,
+          content: 'eslint Success'
+        }]
       })]
     }
   }
@@ -46,7 +53,7 @@ jest.mock('ora', () => {
   return ora
 })
 
-const runDoctor = run('doctor', ['commands/doctor'])
+const runDoctor = run('doctor', [require.resolve('@tarojs/plugin-doctor')])
 
 describe('doctor', () => {
   it('should exit because there isn\'t a Taro project', async () => {
@@ -57,7 +64,6 @@ describe('doctor', () => {
       throw new Error()
     })
     logSpy.mockImplementation(() => {})
-
     try {
       await runDoctor('', { options: { disableGlobalConfig: true } })
     } catch (error) {} // eslint-disable-line no-empty
@@ -66,37 +72,6 @@ describe('doctor', () => {
     expect(logSpy).toBeCalledWith(chalk.red('找不到项目配置文件config/index，请确定当前目录是 Taro 项目根目录!'))
 
     exitSpy.mockRestore()
-    logSpy.mockRestore()
-  })
-
-  it('should log reports', async () => {
-    const NOTE_ALL_RIGHT = chalk.green('[✓] ')
-    const NOTE_VALID = chalk.yellow('[!] ')
-    const NOTE_INVALID = chalk.red('[✗] ')
-
-    const titleChalk = chalk.hex('#aaa')
-    const lineChalk = chalk.hex('#fff')
-    const solutionChalk = chalk.hex('#999')
-
-    const logSpy = jest.spyOn(console, 'log')
-    logSpy.mockImplementation(() => {})
-
-    await runDoctor(path.join(__dirname, 'fixtures/default'))
-
-    expect(logSpy).nthCalledWith(1, '\n' + titleChalk('configValidator'))
-    expect(logSpy).nthCalledWith(2, '  ' + NOTE_INVALID + lineChalk('A'))
-
-    expect(logSpy).nthCalledWith(3, '\n' + titleChalk('packageValidator'))
-    expect(logSpy).nthCalledWith(4, '  ' + NOTE_VALID + lineChalk('B'))
-    expect(logSpy).nthCalledWith(5, '  ' + NOTE_VALID + lineChalk('C'))
-    expect(logSpy).nthCalledWith(6, '      ' + solutionChalk('c'))
-
-    expect(logSpy).nthCalledWith(7, '\n' + titleChalk('recommandValidator'))
-    expect(logSpy).nthCalledWith(8, `  ${NOTE_ALL_RIGHT}没有发现问题`)
-
-    expect(logSpy).nthCalledWith(9, '\n' + titleChalk('eslintValidator'))
-    expect(logSpy).nthCalledWith(10, 'eslint msg')
-
     logSpy.mockRestore()
   })
 })
