@@ -10,7 +10,7 @@ import { serializeParams } from '../../../utils'
 // @ts-ignore
 const { Link } = Taro
 
-function generateRequestUrlWithParams (url: string, params?: unknown) {
+function generateRequestUrlWithParams (url = '', params?: unknown) {
   params = typeof params === 'string' ? params : serializeParams(params)
   if (params) {
     url += (~url.indexOf('?') ? '&' : '?') + params
@@ -19,26 +19,21 @@ function generateRequestUrlWithParams (url: string, params?: unknown) {
   return url
 }
 
-// FIXME 移除 any 标注
-function _request (options) {
-  options = options || {}
-  if (typeof options === 'string') {
-    options = {
-      url: options
-    }
-  }
+function _request (options: Partial<Taro.request.Option> = {}) {
   const { success, complete, fail } = options
-  let url = options.url
-  const params: any = {}
+  let url = options.url || ''
+  const params: RequestInit = {}
   const res: any = {}
   if (options.jsonp) {
-    Object.assign(params, options)
-    params.params = options.data
-    params.cache = options.jsonpCache
-    if (typeof options.jsonp === 'string') {
-      params.name = options.jsonp
+    const { jsonp, ...opts } = options
+    Object.assign(params, opts)
+    // @ts-ignore
+    params.params = opts.data
+    params.cache = opts.jsonpCache
+    if (typeof jsonp === 'string') {
+      // @ts-ignore
+      params.name = jsonp
     }
-    delete params.jsonp
     return jsonpRetry(url, params)
       .then(data => {
         res.statusCode = 200
@@ -149,6 +144,14 @@ function taroInterceptor (chain) {
 
 const link = new Link(taroInterceptor)
 
-export const request: typeof Taro.request = link.request.bind(link)
+export const request = (<T extends Partial<Taro.request.Option> = TaroGeneral.IAnyObject>(...args: [string | T, T]) => {
+  const [url = '', options = {} as T] = args
+  if (typeof url === 'string') {
+    options.url = url
+  } else {
+    Object.assign(options, url)
+  }
+  return link.request(options)
+}) as typeof Taro.request
 export const addInterceptor = link.addInterceptor.bind(link)
 export const cleanInterceptors = link.cleanInterceptors.bind(link)
