@@ -227,14 +227,17 @@ function hideModal () {
   modal.hide()
 }
 
-const showActionSheet: typeof Taro.showActionSheet = async (options = { itemList: [] }) => {
+const showActionSheet = async (
+  options: Taro.showActionSheet.Option = { itemList: [] },
+  methodName = 'showActionSheet'
+): Promise<Taro.showActionSheet.SuccessCallbackResult> => {
   init(document)
   options = Object.assign({
     itemColor: '#000000',
     itemList: []
   }, options)
   const { success, fail, complete } = options
-  const handle = new MethodHandler<Taro.showActionSheet.SuccessCallbackResult>({ name: 'showActionSheet', success, fail, complete })
+  const handle = new MethodHandler<Taro.showActionSheet.SuccessCallbackResult>({ name: methodName, success, fail, complete })
 
   // list item String
   if (!Array.isArray(options.itemList)) {
