@@ -2,10 +2,10 @@
 import { Current, eventCenter, PageInstance, requestAnimationFrame } from '@tarojs/runtime'
 import queryString from 'query-string'
 
-import { loadAnimateStyle } from '../animation'
 import { bindPageResize } from '../events/resize'
 import { bindPageScroll } from '../events/scroll'
-import { setHistoryMode } from '../history'
+import { history,setHistoryMode } from '../history'
+import { loadAnimateStyle, loadRouterStyle } from '../style'
 import { initTabbar } from '../tabbar'
 import { addLeadingSlash, getCurrentPage, getHomePage, routesAlias, stripBasename, stripTrailing } from '../utils'
 import stacks from './stack'
@@ -112,6 +112,17 @@ export default class PageHandler {
     return search.substring(1)
   }
 
+  get usingWindowScroll () {
+    let usingWindowScroll = false
+    if (typeof this.pageConfig?.usingWindowScroll === 'boolean') {
+      usingWindowScroll = this.pageConfig.usingWindowScroll
+    }
+    const win = window as any
+    win.__taroAppConfig ||= {}
+    win.__taroAppConfig.usingWindowScroll = usingWindowScroll
+    return usingWindowScroll
+  }
+
   getQuery (stamp = '', search = '', options: Record<string, unknown> = {}) {
     search = search ? `${search}&${this.search}` : this.search
     const query = search
@@ -124,8 +135,10 @@ export default class PageHandler {
 
   mount () {
     setHistoryMode(this.routerMode, this.router.basename)
+    this.pathname = history.location.pathname
 
     this.animation && loadAnimateStyle(this.animationDuration)
+    loadRouterStyle(this.usingWindowScroll)
 
     const appId = this.appId
     let app = document.getElementById(appId)
@@ -193,7 +206,7 @@ export default class PageHandler {
       this.isTabBar(this.pathname) && pageEl.classList.add('taro_tabbar_page')
       this.addAnimation(pageEl, pageNo === 0)
       page.onShow?.()
-      this.bindPageEvents(page, pageEl, pageConfig)
+      this.bindPageEvents(page, pageConfig)
       this.triggerRouterChange()
     } else {
       page.onLoad?.(param, () => {
@@ -202,7 +215,7 @@ export default class PageHandler {
         this.addAnimation(pageEl, pageNo === 0)
         this.onReady(page, true)
         page.onShow?.()
-        this.bindPageEvents(page, pageEl, pageConfig)
+        this.bindPageEvents(page, pageConfig)
         this.triggerRouterChange()
       })
     }
@@ -249,7 +262,7 @@ export default class PageHandler {
       setDisplay(pageEl)
       this.addAnimation(pageEl, pageNo === 0)
       page.onShow?.()
-      this.bindPageEvents(page, pageEl, pageConfig)
+      this.bindPageEvents(page, pageConfig)
       this.triggerRouterChange()
     } else {
       page.onLoad?.(param, () => {
@@ -257,7 +270,7 @@ export default class PageHandler {
         this.addAnimation(pageEl, pageNo === 0)
         this.onReady(page, false)
         page.onShow?.()
-        this.bindPageEvents(page, pageEl, pageConfig)
+        this.bindPageEvents(page, pageConfig)
         this.triggerRouterChange()
       })
     }
@@ -311,15 +324,18 @@ export default class PageHandler {
       ? document.querySelector(`.taro_page#${id}`)
       : document.querySelector('.taro_page') ||
     document.querySelector('.taro_router')) as HTMLDivElement
-    return el || window
+    return el
   }
 
-  bindPageEvents (page: PageInstance, pageEl?: HTMLElement | null, config: Partial<PageConfig> = {}) {
-    if (!pageEl) {
-      pageEl = this.getPageContainer() as HTMLElement
-    }
+  getScrollingElement (page?: PageInstance | null) {
+    if (this.usingWindowScroll) return window
+    return this.getPageContainer(page) || window
+  }
+
+  bindPageEvents (page: PageInstance, config: Partial<PageConfig> = {}) {
+    const scrollEl = this.getScrollingElement(page)
     const distance = config.onReachBottomDistance || this.config.window?.onReachBottomDistance || 50
-    bindPageScroll(page, pageEl, distance)
+    bindPageScroll(page, scrollEl, distance)
     bindPageResize(page)
   }
 
