@@ -233,30 +233,42 @@ declare module '../../index' {
 
   namespace chooseVideo {
     interface Option {
-      /** 默认拉起的是前置或者后置摄像头。部分 Android 手机下由于系统 ROM 不支持无法生效 */
+      /** 默认拉起的是前置或者后置摄像头。部分 Android 手机下由于系统 ROM 不支持无法生效
+       * @default "back"
+       * @supported weapp, h5
+       */
       camera?: keyof Camera
+      /** 是否压缩所选择的视频文件
+       * @default true
+       * @supported weapp
+       */
+      compressed?: boolean
+      /** 拍摄视频最长拍摄时间，单位秒
+       * @default 60
+       * @supported weapp
+       */
+      maxDuration?: number
+      /** 视频选择的来源
+       * @default ['album', 'camera']
+       * @supported weapp, h5
+       */
+      sourceType?: Array<keyof sourceType>
       /** 接口调用结束的回调函数（调用成功、失败都会执行） */
       complete?: (res: TaroGeneral.CallbackResult) => void
-      /** 是否压缩所选择的视频文件 */
-      compressed?: boolean
       /** 接口调用失败的回调函数 */
       fail?: (res: TaroGeneral.CallbackResult) => void
-      /** 拍摄视频最长拍摄时间，单位秒 */
-      maxDuration?: number
-      /** 视频选择的来源 */
-      sourceType?: Array<keyof sourceType>
       /** 接口调用成功的回调函数 */
       success?: (result: SuccessCallbackResult) => void
     }
     interface SuccessCallbackResult extends TaroGeneral.CallbackResult {
-      /** 选定视频的时间长度 */
-      duration: number
-      /** 返回选定视频的高度 */
-      height: number
-      /** 选定视频的数据量大小 */
-      size: number
       /** 选定视频的临时文件路径 */
       tempFilePath: string
+      /** 选定视频的时间长度 */
+      duration: number
+      /** 选定视频的数据量大小 */
+      size: number
+      /** 返回选定视频的高度 */
+      height: number
       /** 返回选定视频的宽度 */
       width: number
       /** 调用结果 */
@@ -276,30 +288,53 @@ declare module '../../index' {
     }
   }
 
-
   namespace chooseMedia {
     interface Option {
-      /** 最多可以选择的文件个数 */
+      /** 最多可以选择的文件个数
+       * @default 9
+       * @supported weapp, h5
+       */
       count?: number
-      /** 文件类型 */
+      /** 文件类型
+       * @default ['image', 'video']
+       * @supported weapp, h5
+       */
       mediaType?: Array<keyof mediaType>
-      /** 图片和视频选择的来源 */
+      /** 图片和视频选择的来源
+       * @default ['album', 'camera']
+       * @supported weapp, h5
+       */
       sourceType?: Array<keyof sourceType>
-      /** 拍摄视频最长拍摄时间，单位秒。时间范围为 3s 至 30s 之间 */
+      /** 拍摄视频最长拍摄时间，单位秒。时间范围为 3s 至 60s 之间
+       * @default 10
+       * @supported weapp
+       */
       maxDuration?: number
-      /** 仅对 mediaType 为 image 时有效，是否压缩所选文件 */
+      /** 是否压缩所选文件
+       * @default ['original', 'compressed']
+       * @supported weapp
+       */
       sizeType?: Array<'original' | 'compressed'>
-      /** 仅在 sourceType 为 camera 时生效，使用前置或后置摄像头 */
+      /** 仅在 sourceType 为 camera 时生效，使用前置或后置摄像头
+       * @default "back"
+       * @supported weapp, h5
+       */
       camera?: string
+      /** 接口调用结束的回调函数（调用成功、失败都会执行） */
+      complete?: (res: TaroGeneral.CallbackResult) => void
       /** 接口调用失败的回调函数 */
       fail?: (res: TaroGeneral.CallbackResult) => void
       /** 接口调用成功的回调函数 */
       success?: (result: SuccessCallbackResult) => void
+      /** 用来上传的input元素ID
+       * @supported h5
+       */
+      mediaId?: string
     }
     interface SuccessCallbackResult extends TaroGeneral.CallbackResult {
       /** 本地临时文件列表 */
       tempFiles: ChooseMedia[]
-      /** 文件类型，有效值有 image 、video */
+      /** 文件类型，有效值有 image 、video、mix */
       type: string
     }
     /** 本地临时文件列表 */
@@ -316,6 +351,12 @@ declare module '../../index' {
       width: number
       /** 视频缩略图临时文件路径 */
       thumbTempFilePath: string
+      /** 选择的文件的类型 */
+      fileType: string
+      /** 原始的浏览器 File 对象
+       * @supported h5
+       */
+      originalFileObj?: File
     }
     interface mediaType {
       /** 只能拍摄视频或从相册选择视频 */
@@ -350,7 +391,7 @@ declare module '../../index' {
      * @example
      ```tsx
      * Taro.saveVideoToPhotosAlbum({
-     *   filePath: 'wxfile://xxx',
+     *   filePath: 'file://xxx',
      *   success: function (res) {
      *     console.log(res.errMsg)
      *   }
@@ -447,7 +488,7 @@ declare module '../../index' {
     compressVideo(option: compressVideo.Option): Promise<compressVideo.SuccessCallbackResult>
 
     /** 拍摄视频或从手机相册中选视频。
-     * @supported weapp, rn
+     * @supported weapp, h5, rn
      * @example
      * ```tsx
      * Taro.chooseVideo({
@@ -464,7 +505,7 @@ declare module '../../index' {
     chooseVideo(option: chooseVideo.Option): Promise<chooseVideo.SuccessCallbackResult>
 
     /** 拍摄或从手机相册中选择图片或视频。
-     * @supported weapp
+     * @supported weapp, h5
      * @example
      * ```tsx
      * Taro.chooseMedia({
